게시물 수정,삭제 세션 적용 + 테스트 코드 작성

+++ src/main/java/com/board/service/PostService.java
@@ -3,15 +3,20 @@ package com.board.service;
 import com.board.domain.post.Post;
 import com.board.domain.post.PostRepository;
 import com.board.domain.post.PostType;
+import com.board.domain.user.User;
+import com.board.web.HttpSessionUtils;
 import com.board.web.dto.post.PostRequestDto;
 import com.board.web.dto.post.PostResponseDto;
 import com.board.web.dto.post.PostUpdateDto;
 import java.util.stream.Collectors;
+import javax.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class PostService {
@@ -46,15 +51,33 @@ public class PostService {
   }
 
   @Transactional
-  public Long update(Long id, PostUpdateDto postUpdateDto) {
+  public Long update(Long id, PostUpdateDto postUpdateDto, HttpSession session) {
     Post post = postRepository.findById(id)
         .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시물"));
-    post.update(postUpdateDto.getTitle(), postUpdateDto.getContent());
+    if (!HttpSessionUtils.isLoginUser(session)) {
+      throw new IllegalStateException("로그인되지 않은 상태에서 게시물 수정 접근");
+    }
+    User user = HttpSessionUtils.getUserFromSession(session);
+    if (user.matchUserId(postRepository.findById(id).get().getUserId())) {
+      post.update(postUpdateDto.getTitle(), postUpdateDto.getContent());
+      log.info(user.getId() + "님의 id : " + id + "게시글이 업데이트되었습니다.");
+    } else {
+      throw new IllegalStateException("타인의 게시물 삭제 접근");
+    }
     return id;
   }
 
   @Transactional
-  public void delete(Long id){
-    postRepository.deleteById(id);
+  public void delete(Long id, HttpSession session) {
+    if (!HttpSessionUtils.isLoginUser(session)) {
+      throw new IllegalStateException("로그인되지 않은 상태에서 게시물 삭제 접근");
+    }
+    User user = HttpSessionUtils.getUserFromSession(session);
+    if (user.matchUserId(postRepository.findById(id).get().getUserId())) {
+      postRepository.deleteById(id);
+      log.info(user.getId() + "님의 id : " + id + "게시글이 삭제되었습니다.");
+    } else {
+      throw new IllegalStateException("타인의 게시물 삭제 접근");
+    }
   }
 }
